feat(inventories): implement inventory repository methods

+++ app/src/main/java/com/liquotrack/stocksip/features/inventorymanagement/inventories/data/remote/models/InventoryDtoItem.kt
@@ -31,7 +31,7 @@ data class InventoryDtoItem(
     @SerializedName("type")
     val type: String,
     @SerializedName("unitPrice")
-    val unitPrice: Int,
+    val unitPrice: Double,
     @SerializedName("warehouseId")
     val warehouseId: String
 )
